+++ packages/expo-router/src/__tests__/testing-library.test.ios.tsx
@@ -6,30 +6,61 @@ import { renderRouter } from '../testing-library';
 /*
  * Smoke Tests for the Testing Library. While we use these functions in the other tests, we want to make sure they work as expected.
  */
-it('toHavePathname', () => {
-  renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
-  expect(screen).toHavePathname('/home');
+describe('toHavePathname', () => {
+  it('correctly matches', () => {
+    renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
+    expect(screen).toHavePathname('/home');
+  });
+
+  it('fails with the correct message', () => {
+    renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
+    expect(() => expect(screen).toHavePathname('/xyz')).toThrow(/\.toHavePathname/);
+    expect(() => expect(screen).toHavePathname('/xyz')).toThrow(/Expected: "\/xyz"/);
+    expect(() => expect(screen).toHavePathname('/xyz')).toThrow(/Received: "\/home"/);
+  });
+
+  it('fails with the correct message for a .not assertion', () => {
+    renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
+    expect(() => expect(screen).not.toHavePathname('/home')).toThrow(/\.not\.toHavePathname/);
+    expect(() => expect(screen).not.toHavePathname('/home')).toThrow(/Expected: not "\/home"/);
+    expect(() => expect(screen).not.toHavePathname('/home')).toThrow(/Received: "\/home"/);
+  });
 });
 
-it('formats negated matcher messages', () => {
-  renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
-  expect(() => expect(screen).not.toHavePathname('/home')).toThrow(/\.not\.toHavePathname/);
-  expect(() => expect(screen).not.toHavePathname('/home')).toThrow(/Expected: not/);
-});
+describe('toHavePathnameWithParams', () => {
+  it('correctly matches', () => {
+    renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
+    expect(screen).toHavePathnameWithParams('/home?test=true');
+  });
 
-it('toHavePathnameWithParams', () => {
-  renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
-  expect(screen).toHavePathnameWithParams('/home?test=true');
+  it('fails with the correct message', () => {
+    renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
+    expect(() => expect(screen).toHavePathnameWithParams('/xyz')).toThrow(
+      /\.toHavePathnameWithParams/
+    );
+    expect(() => expect(screen).toHavePathnameWithParams('/xyz')).toThrow(/Expected: "\/xyz"/);
+    expect(() => expect(screen).toHavePathnameWithParams('/xyz')).toThrow(
+      /Received: "\/home\?test=true"/
+    );
+  });
+
+  it('fails with the correct message for a .not assertion', () => {
+    renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
+    expect(() => expect(screen).not.toHavePathnameWithParams('/home?test=true')).toThrow(
+      /\.not\.toHavePathnameWithParams/
+    );
+    expect(() => expect(screen).not.toHavePathnameWithParams('/home?test=true')).toThrow(
+      /Expected: not "\/home\?test=true"/
+    );
+    expect(() => expect(screen).not.toHavePathnameWithParams('/home?test=true')).toThrow(
+      /Received: "\/home\?test=true"/
+    );
+  });
 });
 
 it('toHaveSearchParams', () => {
   renderRouter(['[slug]/[...catchAll]'], { initialUrl: '/home/long/name?test=true' });
   expect(screen).toHaveSearchParams({ slug: 'home', test: 'true', catchAll: ['long', 'name'] });
-});
-
-it('toHavePathnameWithParams', () => {
-  renderRouter(['[slug]'], { initialUrl: '/home?test=true' });
-  expect(screen).toHavePathnameWithParams('/home?test=true');
 });
 
 // This test is currently broken in React Navigation v7 as @react-navigation/routers still has the prerenderRoutes key